move connection.py to common
